crates/dc-mini-host-py/pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "dc-mini-host-py"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "rerun-sdk>=0.22.1",
]
[tool.maturin]
features = ["pyo3/extension-module"]
